Make LessonLoader check the Lesson
Create static Helper class which holds the paths to the Lesson directory and to
the content file in this.

Create LessonLoader class which is a QThread so far inspecting whether
the Lesson paths and files exist. Later it should also load and parse
the content. It also checks validity of the xml files.

Add basic draft of XSD to resources.