Pass lesson from LessonLoader to MainWindow
Create signal from LessonLoader and connect it to mainWindow.loaded()
inside of OpenDialog.
Pass the wordList from the LessonLoader in this way to the mainWindow.
I use a QPair for now to also have the Error messages, should look into
Qt Error handling later and maybe use Exceptions.
Get rid of some connections that are not needed and some functions that
were arbitrary.