Display words in the mainWindow
Helper:
Rename getLessonsPath() into getLessonsDirectort() - it gives the main
lesson directory.
Add getLessonPath() it returns the path of a certain lesson.
Make changes in classes that use that already.

Create Lesson class:
The wordList is now prat of a Lesson class, with the name of the lesson.
For now this helps us store the name and words together. Later I might
add statistics and variables to have good learning scheduling.

the loaded() signal from LessonLoader now has the Leson instead of the wordList ask
parameter. The slot (in mainMenu) also changed as did the connect in the
openDialog.

Simple vocabulary view:
When a lesson is loaded in the mainWindow the pictogram gif, the
transliteration and the translation get displayed in the frameLesson.
frameOpening gets hidden.

Give the mainWindow an 'exit' menu entry.