Play audio files
Add variables to store version in Lesson.

LessonLoader:
In Run():
Let loadLesson() return a Lesson instead of a QPair.
This was old Garbage from before having a lesson and just returning the wordList with a QString for errors.

In loadLesson():
Parse the audio tag and save it in the word.
Parse the lesson version and save it in the lesson.

MainWindow:
If Word has an audio enable speaker button and play the file upon clicking.