Simplify loading of lessons
In the last commit I forgot to scan for the 'audio' element with the SAX
parser.

LessonLoader was simplified:
Dont have the lesson name as a member variable but give it to methods.

Make methods which dont change the state const.

Don't run LessonLoader in a Thread for now, will be changed later.

Get rid of signal for lesson loading but pass lesson around.

Get rid of some pointers.