Add Layouts everywhere
Add Layouts everywhere, the 'Example' TextBrowser sometimes didnt show
the last line properly. It turns out this was because it wasn't part of
a Layout and thus Qt had problems calculating its size correctly.
Everything should be part of a Layout it seems.
In acorrdance with that I put everything in layouts.

But this made problems with when you make a widget invisible, as it gets
removedt from the layout tree.
I rely on that when displaing the next 'part' of a word (like
transliteration, translation etc).

It was a bad implementation anyway as view should be devided from state.

So now this is done enum class 'VisisibleWidget'.
An enum class is like an enum but one has namespaces, which is nice.

For this I dont get the word in displayRandomWord and directly set it to all
TextBrowsers. But only get the random word id in chooseRandomWord and store it in a
member variable 'mCurrentWordIndex'.

Then in displayWord I first clear all widgets, and make the relation
between the enum and the widgets using a map.